Add cucumber and SpringRunner to test
