+++ pom.xml
@@ -134,6 +134,25 @@
             <artifactId>jarchivelib</artifactId>
             <version>0.7.1</version>
         </dependency>
+        <!-- Cucumber -->
+        <dependency>
+            <groupId>info.cukes</groupId>
+            <artifactId>cucumber-junit</artifactId>
+            <version>1.2.5</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>info.cukes</groupId>
+            <artifactId>cucumber-java</artifactId>
+            <version>1.2.5</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>info.cukes</groupId>
+            <artifactId>cucumber-spring</artifactId>
+            <version>1.2.5</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
@@ -235,7 +254,7 @@
                               <exclude>**/functionaltest/*</exclude>
                           </excludes>
                           <includes>
-                              <include>**/TestBridgeEndpoints.java</include>
+                              <include>**/TestBridgeEndpointsRunner.java</include>
                           </includes>
                       </configuration>
                   </execution>
