+++ src/integrationtest/java/com/ericsson/ei/frontend/TestBridgeEndpointsRunner.java
@@ -1,0 +1,14 @@
+package com.ericsson.ei.frontend;
+
+import org.junit.runner.RunWith;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(features = "src/integrationtest/resources/features/auth.feature", glue = {
+        "com.ericsson.ei.frontend" }, plugin = {
+                "html:target/cucumber-reports/TestBridgeEndpointsRunner" })
+public class TestBridgeEndpointsRunner {
+
+}
