+++ src/integrationtest/resources/features/auth.feature
@@ -1,0 +1,6 @@
+@BridgeAuthenticationFeature
+Feature: Authentication test
+
+  @AuthenicationLoginScenario
+  Scenario: Authentication login
+    Given frontend is up and running
