+++ src/integrationtest/resources/integration-test.properties
@@ -1,1 +1,1 @@
-ei.it.backend-port:@ei.it.backend-port@
+ei.it.backend-port: @ei.it.backend-port@
